qemu.rb: Open QMP connection on demand
If parsing the command line fails, qemu will not allow a QMP connection.
Trying to connect to it in the constructor will thus make the
constructor fail and prevent users from reading stdout and stderr.

Create the QMP connection on demand so that instead the first .qmp()
access fails and users receive a VM object whose stdout and stderr they
can read.